fix(schema.html): Remove inactive project/site
The listing for 'resumefodder.com' appears to be useless, as the site is
not operational, and the project itself was last updated 3 years ago (at
this repo: https://gitlab.com/steve-perkins/ResumeFodder). Since this is
the case, it should probably be removed.

This change also corrects inconsistent HTML structure (which is also a
result of this listing's existence).